+++ back/src/main/java/com/stock/survive/entity/StockItems.java
@@ -16,7 +16,7 @@ public class StockItems {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "item_no")
-    private Integer itemNo;
+    private Long itemNo;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "item_category_no", nullable = false)
