+++ back/src/main/java/com/stock/survive/repository/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findBySocialEmail(String email);
     boolean existsByNickname(String nickname);
 
@@ -17,7 +17,7 @@ public interface UserRepository extends JpaRepository<User, Integer> {
         left join fetch u.identities oi
         where u.id = :id
     """)
-    Optional<User> findWithIdentitiesById(@org.springframework.data.repository.query.Param("id") Integer id);
+    Optional<User> findWithIdentitiesById(@org.springframework.data.repository.query.Param("id") Long id);
 
     // 마이페이지 상세: 즐겨찾기까지 필요할 때만 fetch join
     @Query("""
@@ -25,11 +25,11 @@ public interface UserRepository extends JpaRepository<User, Integer> {
       left join fetch u.favorites
       where u.id = :id
     """)
-    Optional<User> findWithFavoritesById(@org.springframework.data.repository.query.Param("id") Integer id);
+    Optional<User> findWithFavoritesById(@org.springframework.data.repository.query.Param("id") Long id);
 
     // 메인에서 사용할 것 가볍게
     @Query("select f.itemNo from User u join u.favorites f where u.id = :userId")
-    List<Integer> findFavoriteItemIds(@org.springframework.data.repository.query.Param("userId") Integer userId);
+    List<Long> findFavoriteItemIds(@org.springframework.data.repository.query.Param("userId") Long userId);
 
 
 }
