+++ back/src/main/java/com/stock/survive/service/FavoriteCommandService.java
@@ -2,5 +2,5 @@ package com.stock.survive.service;
 
 public interface FavoriteCommandService {
    // 토글로 관심종목 등록&취소
-    boolean toggleFavorite(Integer userId, /*Short or Integer*/ Integer itemNo);
+    boolean toggleFavorite(Long userId, /*Short or Integer*/ Long itemNo);
 }
